requirements.txt
fastapi
uvicorn[standard]
python-multipart
pillow
numpy
scikit-learn
torch
torchvision
python-dotenv